+++ Assets/Prefabs/JaewonPrefaps/King Slime.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f9d30633fab4da748bc6cf9f7642c744
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
