+++ Assets/Prefabs/JaewonPrefaps/Leaf Slime.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d0c0299dcf8f4e841beb9c20ddccc41f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
