+++ Assets/Prefabs/JaewonPrefaps/Normal Slime.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ce48f6b4a03dd9c4494a6b7916b75b2e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
