+++ Assets/Prefabs/JaewonPrefaps/Rabit Slime.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d936c0d4691d2b24ea3c96e4b1c1f994
+guid: f3673618038fdca4d99fa77d6eb59104
 PrefabImporter:
   externalObjects: {}
   userData:
