+++ Assets/Scripts/SlimeSpawner.cs.meta
